requirement.txt
beautifulsoup4
numpy
Pillow
requests
streamlit
tensorflow
matplotlib
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu
opencv-python==4.5.5.64
gunicorn==19.3.0